requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
autopep8==2.0.4
Django==5.0.2
django-cors-headers==4.3.1
django-redis==5.4.0
django-silk==5.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gprof2dot==2022.7.29
mysqlclient==2.2.4
pillow==10.2.0
pycodestyle==2.11.1
pyenchant==3.2.2
PyJWT==2.8.0
pytz==2024.1
redis==5.0.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
